Set up tableview to accept data, set up chartView, set up notification oberservation
