+++ .swiftlint.yml
@@ -15,7 +15,10 @@ disabled_rules: # rule identifiers to exclude from running
   - todo
   - trailing_comma
   - vertical_parameter_alignment # does not work with tab indentation: https://github.com/realm/SwiftLint/issues/1573
-
+  - trailing_whitespace
+  - line_length
+  - force_try
+  - force_cast
 excluded: # paths to ignore during linting.
   - Carthage
   - Swift Style Guide.playground
